adapt to C99
Strategy being:
- Remove unused variables, unless they are ref'd in some block of
  '#if 0' or commented code, in which case move the decls to inside
  those regions, or comment out.
- Mark any function/block-level function prototypes as extern if not
  already, which I'll use to collate later to decide whether there's
  merit in moving them.

+++ src/demo.h
@@ -52,7 +52,7 @@ extern int DemoRecCnt;                    // Can only record 1-player game
 
 // Demo File - reading from group
 #if DEMO_FILE_TYPE == DEMO_FILE_GROUP
-typedef long DFILE;
+typedef int DFILE;
 #define DREAD(ptr, size, num, handle) kread((handle),(ptr),(size)*(num))
 #define DOPEN_READ(name) kopen4load(name,0)
 #define DCLOSE(handle) kclose(handle)
